requirements.txt
Django==1.10
djangorestframework==3.8.2
Pillow==5.2.0
